+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "start": "node dist/index.js",
     "dev": "NODE_ENV=development nodemon src/index.ts",
-    "build": "tsc --build"
+    "build": "tsc --build",
+    "test": "jest --runInBand --detectOpenHandles --forceExit"
   },
   "author": "",
   "license": "ISC",
@@ -18,7 +19,12 @@
     "typescript": "^5.5.2"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.12",
+    "@types/supertest": "^6.0.2",
+    "jest": "^29.7.0",
     "nodemon": "^3.1.4",
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.1.5",
     "tsx": "^4.15.7"
   }
 }
