+++ src/index.ts
@@ -1,43 +1,11 @@
-import { App } from "./app";
 import { DiscordClient } from "./discord-client";
-import "./env";
 import { logger } from "./logger";
+import app from "./app";
+import "./env";
 
 logger.init("Initializing server...");
 
 const port = process.env.PORT || 4000;
-
-const app = new App();
-
-app.GET(
-  "/",
-  (req, res) => {
-    res.send("App is running");
-  },
-  true
-);
-
-app.GET("/overlay", async (req, res) => {
-  res.json(app.overlay);
-});
-
-app.POST("/overlay", async (req, res) => {
-  try {
-    const data = await req.json();
-    app.updateOverlay(data);
-    logger.info("Overlay updated");
-  } catch (err) {
-    logger.warn("Error updating overlay:", err);
-    res.send("Invalid overlay data", 400);
-    return;
-  }
-  res.send("Overlay updated");
-});
-
-// Handle connections at /overlay namespace
-app.io.of("/overlay").on("connection", (socket) => {
-  socket.emit("overlay", app.overlay);
-});
 
 // Start server
 app.listen(port, () => {
